+++ docs/.vitepress/en/config.js
@@ -1,0 +1,12 @@
+import { navbar } from "./navbar";
+import { sidebar } from "./sidebar";
+
+export const en = {
+    label: "English",
+    lang: "en",
+    link: "/en/",
+    themeConfig: {
+        nav: navbar,
+        sidebar: sidebar,
+    },
+}
