--- docs/.vitepress/navbar/en.js
@@ -1,41 +1,0 @@
-export const navbar = [
-    {
-      text: "Basic setup",
-      link: "/en/basic/install-launch-server",
-    },
-    {
-      text: "For developers",
-      items: [
-        {
-          text: "Aurora API",
-          items: [
-            {
-              text: "General information",
-              link: "/en/for-developers/api/info",
-            },
-            {
-              text: "List of error codes",
-              link: "/en/for-developers/api/errors",
-            },
-          ],
-        },
-        {
-          text: "Other",
-          items: [
-            {
-              text: "Alternative mirrors",
-              link: "/en/for-developers/mirrors",
-            },
-            {
-              text: "Requests for JSON authorization",
-              link: "/en/for-developers/json-auth",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      text: "Main WebSite",
-      link: "https://aurora-launcher.ru",
-    },
-];
