+++ docs/.vitepress/search.js
@@ -1,0 +1,11 @@
+import { search as ru } from "./ru/config"
+
+export const search = {
+    provider: "local",
+    options: {
+        detailedView: true,
+        locales: {
+          root: ru
+        },
+    },
+}
