--- docs/.vitepress/sidebar/en.js
@@ -1,105 +1,0 @@
-export const sidebar = {
-    "/en/basic/": [
-      {
-        text: "Dictionary of definitions",
-        link: "/en/basic/dictionary",
-      },
-      {
-        text: "Setting Launcher Server",
-        items: [
-          {
-            text: "Getting started",
-            link: "/en/basic/install-launch-server",
-          },
-          {
-            text: "Game clients",
-            link: "/en/basic/clients",
-          },
-          {
-            text: "Setting up security",
-            link: "/en/basic/guard",
-          },
-          {
-            text: "Setting up auth",
-            link: "/en/basic/auth",
-          },
-          {
-            text: "Setting up skins system",
-            link: "/en/basic/system-skin",
-          },
-          {
-            text: "Binding Minecraft server",
-            link: "/en/basic/server",
-          },
-        ],
-      },
-      {
-        text: "Setting up Launcher",
-        items: [
-          {
-            text: "Getting started",
-            link: "/en/basic/install-launch",
-          },
-          {
-            text: "Setting up Discord activity",
-            link: "/en/basic/discord-rpc",
-          },
-          {
-            text: "Setting up build",
-            link: "/en/basic/build",
-          },
-        ],
-      },
-      {
-        text: "Extra setting up",
-        items: [
-          {
-            text: "Launcher signing",
-            link: "/en/basic/signing",
-          },
-          {
-            text: "Setting up Nginx",
-            link: "/en/basic/nginx",
-          },
-          {
-            text: "Setting up Сloudflare",
-            link: "/en/basic/cloudflare",
-          },
-          {
-            text: "Run as a service",
-            link: "/en/basic/demon",
-          },
-        ],
-      },
-    ],
-    "/en/for-developers/api/": [
-      {
-        text: "Aurora API",
-        items: [
-          {
-            text: "General information",
-            link: "/en/for-developers/api/info",
-          },
-          {
-            text: "List of error codes",
-            link: "/en/for-developers/api/errors",
-          },
-        ],
-      },
-    ],
-    "/en/for-developers/": [
-      {
-        text: "Miscellaneous",
-        items: [
-          {
-            text: "Alternative mirrors",
-            link: "/en/for-developers/mirrors",
-          },
-          {
-            text: "Requests for JSON authorization",
-            link: "/en/for-developers/json-auth",
-          },
-        ],
-      },
-    ],
-};
