+++ docs/en/basic/nginx.md
@@ -79,7 +79,7 @@ server {
     }
 }
 ```
-```nginx [no domain]
+```nginx [IPv4]
 upstream ws-launcher {
     server 127.0.0.1:1370;
 }
