+++ docs/ru/basic/demon.md
@@ -32,10 +32,6 @@ pm2 save
 
 ## Запуск через `systemd`
 
-Создаём файл конфигурации сервиса и при необходимости меняем настройки:
-```sh
-sudo nano /etc/systemd/system/launcher.service
-```
 Создаём отдельного юзера для управления Launcher Server:
 ```sh
 adduser launchserver --system --group
@@ -43,6 +39,10 @@ adduser launchserver --system --group
 Чтобы все работало корректно, вам потребуется выдать юзеру права на папку:
 ```sh
 sudo chown launchserver:launchserver Путь до вашей папки с Launcher Server
+```
+Создаём файл конфигурации сервиса и при необходимости меняем настройки:
+```sh
+sudo nano /etc/systemd/system/launcher.service
 ```
 
 Само содержимое файла:
